details

+++ webelexis-server/src/main/java/ch/webelexis/Runner.java
@@ -11,6 +11,12 @@ import java.util.logging.Logger;
 
 /**
  * Created by gerry on 27.06.15.
+ * (1) load default configuration
+ * (2) try to find local configuraton:
+ * (2a) filename given as parameter
+ * (2b) file "cfglocal.json" in current directoy
+ * (3) Merge local configuration into default configuration
+ * (4) Launch Webelexis CoreVerticle with resulting configuration
  */
 public class Runner {
   public static void main(String[] args) {
@@ -18,6 +24,7 @@ public class Runner {
       Logger log = Logger.getGlobal();
       Vertx vertx = Vertx.vertx();
       JsonObject cfg = new JsonObject();
+      // load default configuration
       InputStream in = Runner.class.getResourceAsStream("../../config_defaults.json");
       if (in != null) {
         System.out.println("Running as files");
@@ -32,18 +39,26 @@ public class Runner {
 
         }
       }
-      if ((args.length == 0) || (!new File(args[0]).exists())) {
-        String enc = cfg.encodePrettily();
-        FileWriter out = new FileWriter("cfglocal.json");
-        out.write(enc);
-        out.close();
-
-      } else {
-
+      // check if local config was given as argument
+      if ((args.length > 0) && (new File(args[0]).exists())) {
         JsonObject cfglocal = Cleaner.createFromFile(args[0]);
         cfg = cfg.mergeIn(cfglocal);
 
       }
+      // Check if local configuration file exists
+      else if (new File("cfglocal.json").exists()) {
+        cfg = cfg.mergeIn(Cleaner.createFromFile("cfglocal.json"));
+
+      } else {
+        //  create local configuration file from default configuration
+        String enc = cfg.encodePrettily();
+        FileWriter out = new FileWriter("cfglocal.json");
+        out.write(enc);
+        out.close();
+        System.out.println("Configuration written. Please edit cfglocal.json and restart");
+        System.exit(0);
+      }
+
       vertx.deployVerticle("ch.webelexis.CoreVerticle", new DeploymentOptions().setConfig(cfg));
     } catch (Exception ex) {
       System.out.print("usage: java -jar webelexis [<config>]\\n" + ex.getMessage());
